Fixed the build script to include Unmanaged

+++ build/build.sh
@@ -1,9 +1,11 @@
 #!/usr/bin/bash
+
+DOTNET_VERSION="net9.0"
 FR_VERSION=$(cat version.txt)
 echo "FortRise Version: $FR_VERSION"
 MAIN=$(pwd)
 
-INTERNALS_DIR="FortLauncher/bin/Debug/net9.0/Internals"
+INTERNALS_DIR="FortLauncher/bin/Debug/$DOTNET_VERSION/Internals"
 INTERNALS_DEST_DIR="artifacts/Internals"
 
 if [ -d "$INTERNALS_DIR" ]; then
@@ -12,14 +14,32 @@ if [ -d "$INTERNALS_DIR" ]; then
 fi
 
 echo "Cleaning all binaries of Internal mods"
-rm -r InternalMods/FortRise.Content/bin
-rm -r InternalMods/FortRise.ImGui/bin
+if [ -d "InternalMods/FortRise.Content/bin" ]; then
+    rm -r InternalMods/FortRise.Content/bin
+    rm -r InternalMods/FortRise.ImGui/bin
+fi
 
 echo "Rebuilding all projects for Internals"
 dotnet build -c Release
 
+if [ $? -eq 1 ]; then
+    echo "Build failed"
+    exit 1
+fi
+
+sleep 0.5s
+
 echo "Copying all Internals to artifacts"
 cp -r "$INTERNALS_DIR" "artifacts"
+
+if [ -d "artifacts/FortRise.v$FR_VERSION-win-x64" ]; then
+    echo "Removing all similar artifacts"
+    rm -r artifacts/FortRise.v$FR_VERSION-win-x64
+    rm -r artifacts/FortRise.v$FR_VERSION-linux-x64
+
+    rm -r artifacts/FortRise.v$FR_VERSION-win-x64.zip
+    rm -r artifacts/FortRise.v$FR_VERSION-linux-x64.zip
+fi
 
 echo "Publishing Artifacts for Linux"
 dotnet publish -c Release -r linux-x64 --self-contained
@@ -34,6 +54,16 @@ cp -r "$INTERNALS_DEST_DIR" "artifacts/FortRise.v$FR_VERSION-win-x64/FortRise"
 cp -r "$INTERNALS_DEST_DIR" "artifacts/FortRise.v$FR_VERSION-linux-x64/FortRise"
 # todo macos
 
+mkdir artifacts/FortRise.v$FR_VERSION-linux-x64/FortRise/Internals/FortRise.ImGui/Unmanaged
+mkdir artifacts/FortRise.v$FR_VERSION-win-x64/FortRise/Internals/FortRise.ImGui/Unmanaged
+
+# things that we don't care about yet
+rm -r InternalMods/FortRise.ImGui/bin/Release/$DOTNET_VERSION/runtimes/win-arm64
+rm -r InternalMods/FortRise.ImGui/bin/Release/$DOTNET_VERSION/runtimes/win-x86
+
+cp -r InternalMods/FortRise.ImGui/bin/Release/$DOTNET_VERSION/runtimes artifacts/FortRise.v$FR_VERSION-linux-x64/FortRise/Internals/FortRise.ImGui/Unmanaged
+cp -r InternalMods/FortRise.ImGui/bin/Release/$DOTNET_VERSION/runtimes artifacts/FortRise.v$FR_VERSION-win-x64/FortRise/Internals/FortRise.ImGui/Unmanaged
+
 echo "Zipping all artifacts"
 
 cd "artifacts/FortRise.v$FR_VERSION-win-x64/"
@@ -44,10 +74,8 @@ cd "artifacts/FortRise.v$FR_VERSION-linux-x64/"
 zip -r "../FortRise.v$FR_VERSION-linux-x64.zip" "FortRise"
 cd "$MAIN"
 # todo macos
-
 echo "Cleaning up"
 rm -r "$INTERNALS_DEST_DIR"
 rm -r "$INTERNALS_DIR"
 
-# rerun the build script so I can still run the modloader afterwards
 dotnet build -c Release
